# packages/studio/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "InkOS Studio API",
    "version": "1.5.0",
    "description": "Auto-generated OpenAPI spec from 157 endpoints across 33 route modules."
  },
  "servers": [
    {
      "url": "http://localhost:4579",
      "description": "InkOS Studio API server"
    }
  ],
  "paths": {
...
    "/api/v1/style/authors": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors",
        "operationId": "get__api_v1_style_authors",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors",
        "operationId": "post__api_v1_style_authors",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors/:authorId",
        "operationId": "get__api_v1_style_authors_authorId",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "tags": [
          "authors"
        ],
        "summary": "DELETE /api/v1/style/authors/:authorId",
        "operationId": "delete__api_v1_style_authors_authorId",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/sources": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/:authorId/sources",
        "operationId": "post__api_v1_style_authors_authorId_sources",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/reanalyze": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/:authorId/reanalyze",
        "operationId": "post__api_v1_style_authors_authorId_reanalyze",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/diagnostics": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/:authorId/diagnostics",
        "operationId": "post__api_v1_style_authors_authorId_diagnostics",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors/:authorId/diagnostics",
        "operationId": "get__api_v1_style_authors_authorId_diagnostics",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/diagnostics/{diagnosticsId}": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors/:authorId/diagnostics/:diagnosticsId",
        "operationId": "get__api_v1_style_authors_authorId_diagnostics_diagnosticsId",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/books/{id}/style/apply-author": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/books/:id/style/apply-author",
        "operationId": "post__api_v1_books_id_style_apply_author",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/search": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/search",
        "operationId": "post__api_v1_style_authors_search",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/fetch": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/fetch",
        "operationId": "post__api_v1_style_authors_fetch",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/samples/write": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/samples/write",
        "operationId": "post__api_v1_style_authors_samples_write",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/distillations": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/:authorId/distillations",
        "operationId": "post__api_v1_style_authors_authorId_distillations",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/distillations/current": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors/:authorId/distillations/current",
        "operationId": "get__api_v1_style_authors_authorId_distillations_current",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "patch": {
        "tags": [
          "authors"
        ],
        "summary": "PATCH /api/v1/style/authors/:authorId/distillations/current",
        "operationId": "patch__api_v1_style_authors_authorId_distillations_current",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/distillations/current/publish": {
      "post": {
        "tags": [
          "authors"
        ],
        "summary": "POST /api/v1/style/authors/:authorId/distillations/current/publish",
        "operationId": "post__api_v1_style_authors_authorId_distillations_current_publish",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/style/authors/{authorId}/distillations/versions": {
      "get": {
        "tags": [
          "authors"
        ],
        "summary": "GET /api/v1/style/authors/:authorId/distillations/versions",
        "operationId": "get__api_v1_style_authors_authorId_distillations_versions",
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
...
}